pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "trafilatura>=1.12.2",
    "transformers>=4.45.1",
    "datasets>=3.0.1",
    "requests>=2.32.3",
    "gitpython>=3.1.43",
    "torch",
    "python-dotenv>=1.0.1",
    "beautifulsoup4>=4.12.3",
]
